# src/internal/responses/responses.go
/*
Date Created		14 September 2024
Author				Mike Z
Email				mzinyoni7@outlook.com
Website				https://mikeio.web.app
Status				Looking for a job!
Description			Users and rooms simulation
*/

package responses

type Error400Response struct {
	Error interface{} `json:"error"`
}